Close files to prevent resource leakage
These two files open files to parse them and do not explicitly close
them.  The result is that on machines with large amounts of memory
the heap just continues to grow without GC kicking in to close the
files.  This change explicitly closes the files such that the resource
leakage doesn't happen.

Change-Id: I9f72f027f0b328db624f982e577b1d47b25ae0d4